+++ includes/abstracts/llms.abstract.database.store.php
@@ -4,10 +4,14 @@ defined( 'ABSPATH' ) || exit;
 /**
  * WPDB database interactions
  * @since    3.14.0
- * @version  3.21.0
+ * @version  [version]
  */
 abstract class LLMS_Abstract_Database_Store {
 
+	/**
+	 * The Database ID of the record
+	 * @var  null
+	 */
 	protected $id = null;
 
 	/**
@@ -44,6 +48,14 @@ abstract class LLMS_Abstract_Database_Store {
 	 * @var  string
 	 */
 	protected $table_prefix = 'lifterlms_';
+
+	/**
+	 * The record type
+	 * Used for filters/actions
+	 * Should be defined by extending classes
+	 * @var  string
+	 */
+	protected $type = '';
 
 	/**
 	 * Constructor
@@ -175,7 +187,7 @@ abstract class LLMS_Abstract_Database_Store {
 	 * Create the item in the database
 	 * @return   int|false
 	 * @since    3.14.0
-	 * @version  3.14.0
+	 * @version  [version]
 	 */
 	private function create() {
 
@@ -187,6 +199,8 @@ abstract class LLMS_Abstract_Database_Store {
 		$format = array_map( array( $this, 'get_column_format' ), array_keys( $this->data ) );
 		$res = $wpdb->insert( $this->get_table(), $this->data, $format );
 		if ( 1 === $res ) {
+			$this->id = $wpdb->insert_id;
+			do_action( 'llms_' . $this->type . '_created', $this->id, $this );
 			return $wpdb->insert_id;
 		}
 		return false;
@@ -197,7 +211,7 @@ abstract class LLMS_Abstract_Database_Store {
 	 * Delete the object from the database
 	 * @return   boolean     true on success, false otherwise
 	 * @since    3.14.0
-	 * @version  3.14.0
+	 * @version  [version]
 	 */
 	public function delete() {
 
@@ -205,12 +219,14 @@ abstract class LLMS_Abstract_Database_Store {
 			return false;
 		}
 
+		$id = $this->id;
 		global $wpdb;
 		$where = array_combine( array_keys( $this->primary_key ), array( $this->id ) );
 		$res = $wpdb->delete( $this->get_table(), $where, array_values( $this->primary_key ) );
 		if ( $res ) {
 			$this->id = null;
 			$this->data = array();
+			do_action( 'llms_' . $this->type . '_deleted', $id, $this );
 			return true;
 		}
 		return false;
@@ -248,7 +264,11 @@ abstract class LLMS_Abstract_Database_Store {
 		$format = array_map( array( $this, 'get_column_format' ), array_keys( $data ) );
 		$where = array_combine( array_keys( $this->primary_key ), array( $this->id ) );
 		$res = $wpdb->update( $this->get_table(), $data, $where, $format, array_values( $this->primary_key ) );
-		return $res ? true : false;
+		if ( $res ) {
+			do_action( 'llms_' . $this->type . '_updated', $this->id, $this );
+			return true;
+		}
+		return false;
 
 	}
 
@@ -276,7 +296,7 @@ abstract class LLMS_Abstract_Database_Store {
 	 * Creates is it doesn't already exist, updates if it does
 	 * @return   boolean
 	 * @since    3.14.0
-	 * @version  3.14.0
+	 * @version  [version]
 	 */
 	public function save() {
 
@@ -284,7 +304,6 @@ abstract class LLMS_Abstract_Database_Store {
 
 			$id = $this->create();
 			if ( $id ) {
-				$this->id = $id;
 				return true;
 			}
 			return false;